moved ProdAndRelease component to separate folder
