+++ src/Tile/index.js
@@ -1,21 +1,10 @@
-import {
-  Container,
-  Description,
-  Content,
-  Poster,
-  ProdAndReleaseInfo,
-  Production,
-  ReleaseDateText,
-  ReleaseDateContent,
-  ReleaseDate,
-  ProductionText,
-  ProductionContent,
-} from "./styled";
-import noPoster from "../common/images/no-poster.png";
+import { Container, Description, Content, Poster } from "./styled";
 import { TagsComponent } from "./Tags";
 import { Information } from "./Information";
 import { RaitingComponent } from "./Raiting";
 import { DescriptionComponent } from "./Description";
+import { ProductionAndRelease } from "./ProductionAndRelease";
+import noPoster from "../common/images/no-poster.png";
 
 export const Tile = ({ movieDetails, srcImage }) => {
   return (
@@ -24,25 +13,12 @@ export const Tile = ({ movieDetails, srcImage }) => {
       <Content movieDetails={movieDetails}>
         <Description movieDetails={movieDetails}>
           <Information />
-          {movieDetails ? (
-            <ProdAndReleaseInfo>
-              <Production>
-                <ProductionText>Production:</ProductionText>
-                <ProductionContent>China, USA</ProductionContent>
-              </Production>
-              <ReleaseDate>
-                <ReleaseDateText>Release date:</ReleaseDateText>
-                <ReleaseDateContent>24.10.2020</ReleaseDateContent>
-              </ReleaseDate>
-            </ProdAndReleaseInfo>
-          ) : null}
+          {movieDetails ? <ProductionAndRelease /> : null}
           <TagsComponent movieDetails={movieDetails} />
         </Description>
         <RaitingComponent movieDetails={movieDetails} />
       </Content>
-      {movieDetails ? (
-        <DescriptionComponent />
-      ) : null}
+      {movieDetails ? <DescriptionComponent /> : null}
     </Container>
   );
 };
